sar_agent/requirements.txt
# --- Core framework ---
fastapi==0.115.0
uvicorn==0.30.0

# --- LLM / NLP ---
transformers==4.44.2
torch>=2.3.0
accelerate==0.33.0
sentencepiece
einops
bitsandbytes

# --- Financial LLM (FinGPT) ---
fingpt @ git+https://github.com/AI4Finance-Foundation/FinGPT.git

# --- File parsing ---
pdfplumber==0.11.0       # Extract text from PDFs
python-docx==1.1.2       # Read Word docs
openpyxl==3.1.5          # Excel (XLSX) support
pandas==2.2.3            # CSV + tabular handling
lxml==5.3.0              # XML schema parsing/validation

# --- Utilities ---
requests==2.32.3
python-multipart==0.0.9   # File uploads with FastAPI
pydantic==2.9.2           # Data validation
loguru==0.7.2             # Logging

# --- Testing ---
pytest==8.3.2
httpx==0.27.2

# --- Deployment ---
docker==7.1.0             # Podman/Docker compatibility

fastapi
uvicorn
transformers
torch
pandas
python-docx
pymupdf
reportlab
